Move shared components of docs into one folder
