+++ packages/docs/stories/Welcome.mdx
@@ -1,6 +1,6 @@
 import { Meta } from "@storybook/blocks";
 
-import { Chapter, SymbolLine } from '../src/Chapter.tsx';
+import { Chapter, SymbolLine } from '../src/shared-components/Chapter.tsx';
 
 <Meta title="Welcome" />
