+++ packages/docs/stories/packages/components/5-Changelog.mdx
@@ -1,6 +1,6 @@
 import { Meta } from "@storybook/blocks";
 import Changelog from '../../../../components/CHANGELOG.md?raw';
-import { ChangelogFormatter } from '../../../src/Changelog.tsx'
+import { ChangelogFormatter } from '../../../src/shared-components/Changelog.tsx';
 
 <Meta title="Packages/Components/Changelog" />
