+++ packages/docs/stories/tokens/Colors.mdx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/blocks";
-import { CopyToClipBoard } from '../../src/Clipboard.tsx';
-import { CopyRawHexValue } from '../../src/RawColorValue.tsx';
-import { ColorSwatch, MultiLineDescription } from '../../src/TokenHelpers.tsx';
+import { CopyToClipBoard } from '../../src/shared-components/Clipboard.tsx';
+import { CopyRawHexValue } from '../../src/shared-components/RawColorValue.tsx';
+import { ColorSwatch, MultiLineDescription } from '../../src/shared-components/TokenHelpers.tsx';
 import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
 
 import {
