+++ packages/docs/stories/tokens/Shadows.mdx
@@ -1,8 +1,8 @@
 import { Meta } from "@storybook/blocks";
-import { CopyToClipBoard } from '../../src/Clipboard.tsx';
-import { CopyRawCssPropValue } from '../../src/RawColorValue.tsx';
+import { CopyToClipBoard } from '../../src/shared-components/Clipboard.tsx';
+import { CopyRawCssPropValue } from '../../src/shared-components/RawColorValue.tsx';
 import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { Shadow, MultiLineDescription } from '../../src/TokenHelpers.tsx';
+import { Shadow, MultiLineDescription } from '../../src/shared-components/TokenHelpers.tsx';
 
 import {
   getCSSToken,
