+++ packages/docs/stories/tokens/Typography.mdx
@@ -1,9 +1,9 @@
 import { Meta } from "@storybook/blocks";
 import { camelCase } from 'change-case';
-import { CopyToClipBoard } from '../../src/Clipboard.tsx';
-import { CopyRawCssPropValue , CopyRawValue} from '../../src/RawColorValue.tsx';
+import { CopyToClipBoard } from '../../src/shared-components/Clipboard.tsx';
+import { CopyRawCssPropValue , CopyRawValue} from '../../src/shared-components/RawColorValue.tsx';
 import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { MultiLineDescription } from '../../src/TokenHelpers.tsx';
+import { MultiLineDescription } from '../../src/shared-components/TokenHelpers.tsx';
 import {
   getCSSToken,
   getFontFamily,
